+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 - Modern JavaScript From The Beginning (Udemy)
   - Section 2: JavaScript Language Fundamentals
     - String Methods & Concatenation
+    - Template Literals
 - An example with strings and string methods for String Methods & Concatenation was added.
+- An example with template literals for Template Literals was added.
 
 ## August 3, 2020
 - Modern JavaScript From The Beginning (Udemy)